moves causes to an Interests folder and udpates path in routes
